Eliminada dependencia de Clase Control
El archivo control.py ahora solo tiene un parser de los viejos archivos
.ctl. Devuelve un diccionario igual que si leyera un toml.

Por ahora lo dejamos como documentación mas que nada, porque ahí están
implícitamente registradas la mayoría de las opciones que hay que tener
en cuenta.

También se modificó observations, pero todavía no lo eliminamos porque
se requiere para que el programa pueda seguir escribiendo los archivos
de salida en el formato del programa original (esto por ahora no lo
borramos para no romper todo pero no es necesario).

Se comentó un test inutil en test_noise_creation.

Se agregó un script de bash para chequear que el programa corre sobre
los archivos ctl de valid_ctl. (muy trogloditamente).

Queda bastante trabajo con los parámetros todavía.
